[SPARK-37614][SQL] Support ANSI Aggregate Function: regr_avgx & regr_avgy
### What changes were proposed in this pull request?
`REGR_AVGX` and `REGR_AVGY` are ANSI aggregate functions.

`REGR_AVGX` returns the mean of the independent_variable_expression for all non-null data pairs of the dependent and independent variable arguments.
**Syntax**: `REGR_AVGX(dependent_variable_expression, independent_variable_expression)`
The equation for computing `REGR_AVGX` is: `REGR_AVGX = SUM(x)/n`

**Examples**:
```
> SELECT _FUNC_(y, x) FROM VALUES (1, 2), (2, 2), (2, 3), (2, 4) AS tab(y, x);
  2.75
> SELECT _FUNC_(y, x) FROM VALUES (1, 2), (2, null), (2, 3), (2, 4) AS tab(y, x);
  3.0
> SELECT _FUNC_(y, x) FROM VALUES (1, 2), (2, null), (null, 3), (2, 4) AS tab(y, x);
  3.0
```

`REGR_AVGY` returns the mean of the dependent_variable_expression for all non-null data pairs of the dependent and independent variable arguments.
**Syntax**: `REGR_AVGY(dependent_variable_expression, independent_variable_expression)`
The equation for computing `REGR_AVGY ` is: `REGR_AVGY = SUM(y)/n`

**Examples**:
```
> SELECT _FUNC_(y, x) FROM VALUES (1, 2), (2, 2), (2, 3), (2, 4) AS tab(y, x);
  1.75
> SELECT _FUNC_(y, x) FROM VALUES (1, 2), (2, null), (2, 3), (2, 4) AS tab(y, x);
  1.6666666666666667
> SELECT _FUNC_(y, x) FROM VALUES (1, 2), (2, null), (null, 3), (2, 4) AS tab(y, x);
  1.5
```

**dependent_variable_expression**: the dependent variable for the regression. A dependent variable is something that is measured in response to a treatment. The expression cannot contain any ordered analytical or aggregate functions.

**independent_variable_expression**: the independent variable for the regression. An independent variable is a treatment: something that is varied under your control to test the behavior of another variable.

The expression cannot contain any ordered analytical or aggregate functions.

The mainstream database supports `REGR_AVGX` and `REGR_AVGY` show below:
**Teradata**
https://docs.teradata.com/r/kmuOwjp1zEYg98JsB8fu_A/KkJgUSq2O6JRU3bCK~0cug
**Snowflake**
https://docs.snowflake.com/en/sql-reference/functions/regr_avgx.html
**Oracle**
https://docs.oracle.com/en/database/oracle/oracle-database/19/sqlrf/REGR_-Linear-Regression-Functions.html#GUID-A675B68F-2A88-4843-BE2C-FCDE9C65F9A9
**Vertica**
https://www.vertica.com/docs/9.2.x/HTML/Content/Authoring/SQLReferenceManual/Functions/Aggregate/REGR_AVGX.htm?tocpath=SQL%20Reference%20Manual%7CSQL%20Functions%7CAggregate%20Functions%7C_____24
**DB2**
https://www.ibm.com/docs/en/db2/11.5?topic=af-regression-functions-regr-avgx-regr-avgy-regr-count
**H2**
http://www.h2database.com/html/functions-aggregate.html#regr_avgx
**Postgresql**
https://www.postgresql.org/docs/8.4/functions-aggregate.html
**Sybase**
https://infocenter.sybase.com/help/index.jsp?topic=/com.sybase.help.sqlanywhere.12.0.0/dbreference/regr-avgx-function.html
**Exasol**
https://docs.exasol.com/sql_references/functions/alphabeticallistfunctions/regr_function.htm

### Why are the changes needed?
`REGR_AVGX` and `REGR_AVGY` are very useful.

### Does this PR introduce _any_ user-facing change?
'Yes'. New feature.

### How was this patch tested?
New tests.

Closes #34868 from beliefer/SPARK-37614.

Authored-by: Jiaan Geng <beliefer@163.com>
