+++ sql/core/src/test/resources/sql-tests/results/postgreSQL/aggregates_part1.sql.out
@@ -1,5 +1,5 @@
 -- Automatically generated by SQLQueryTestSuite
--- Number of queries: 45
+-- Number of queries: 46
 
 
 -- !query
@@ -294,6 +294,14 @@ SELECT regr_count(b, a) FROM aggtest
 struct<regr_count(b, a):bigint>
 -- !query output
 4
+
+
+-- !query
+SELECT regr_avgx(b, a), regr_avgy(b, a) FROM aggtest
+-- !query schema
+struct<regr_avgx(b, a):double,regr_avgy(b, a):double>
+-- !query output
+49.5	107.94315227307379
 
 
 -- !query
